// libs/enums/service.enum.ts
export enum ServiceType {
  DAY_CARE = "DAY_CARE",
  WALKING = "WALKING",
  GROOMING = "GROOMING",
  BOARDING = "BOARDING",
  TRAINING = "TRAINING",
  VETERINARY = "VETERINARY",
}

export enum ServiceStatus {
  ACTIVE = "ACTIVE",
  PAUSE = "PAUSE",
  DELETE = "DELETE",
}

export enum ServiceLocation {
  SEOUL = "SEOUL",
  BUSAN = "BUSAN",
  INCHEON = "INCHEON",
  DAEGU = "DAEGU",
  SUWON = "SUWON",
  GYEONGJU = "GYEONGJU",
  GWANGJU = "GWANGJU",
  JEONJU = "JEONJU",
  DAEJEON = "DAEJEON",
  JEJU = "JEJU",
}